backend Product 18-09-2018

+++ src/main/java/com/winstgenerator/backend/DTO/ProductDto.java
@@ -12,7 +12,7 @@ public class ProductDto {
 	
 	//public List<Double> pricesOnline; 
 	
-	private Integer winstMargin;
+	private Double winstMargin;
 
 	public String getProductName() {
 		return productName;
@@ -46,11 +46,11 @@ public class ProductDto {
 	//	this.pricesOnline = pricesOnline;
 	//}
 
-	public Integer getWinstMargin() {
+	public Double getWinstMargin() {
 		return winstMargin;
 	}
 
-	public void setWinstMargin(Integer winstMargin) {
+	public void setWinstMargin(Double winstMargin) {
 		this.winstMargin = winstMargin;
 	}
